+++ src/providers/LocalVideoProvider/index.tsx
@@ -1,21 +1,19 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { AudioVideoObserver, VideoTileState } from 'amazon-chime-sdk-js';
 import React, {
   createContext,
-  useState,
-  useEffect,
-  useContext,
   useCallback,
+  useContext,
+  useEffect,
   useMemo,
+  useState,
 } from 'react';
-import { AudioVideoObserver, VideoTileState } from 'amazon-chime-sdk-js';
 
-import { useMeetingManager } from '../MeetingProvider';
-import { useAudioVideo } from '../AudioVideoProvider';
-
-import { videoInputSelectionToDevice } from '../../utils/device-utils';
 import { LocalVideoContextType } from '../../types';
+import { useAudioVideo } from '../AudioVideoProvider';
+import { useMeetingManager } from '../MeetingProvider';
 
 const Context = createContext<LocalVideoContextType | null>(null);
 
@@ -40,17 +38,21 @@ const LocalVideoProvider: React.FC = ({ children }) => {
   }, [audioVideo]);
 
   const toggleVideo = useCallback(async (): Promise<void> => {
-    if (isVideoEnabled || !meetingManager.selectedVideoInputDevice) {
+    if (isVideoEnabled || !meetingManager.selectedVideoInputTransformDevice) {
       audioVideo?.stopLocalVideoTile();
       setIsVideoEnabled(false);
     } else {
       await audioVideo?.chooseVideoInputDevice(
-        videoInputSelectionToDevice(meetingManager.selectedVideoInputDevice)
+        meetingManager.selectedVideoInputTransformDevice
       );
       audioVideo?.startLocalVideoTile();
       setIsVideoEnabled(true);
     }
-  }, [audioVideo, isVideoEnabled, meetingManager.selectedVideoInputDevice]);
+  }, [
+    audioVideo,
+    isVideoEnabled,
+    meetingManager.selectedVideoInputTransformDevice,
+  ]);
 
   useEffect(() => {
     if (!audioVideo) {
@@ -75,12 +77,10 @@ const LocalVideoProvider: React.FC = ({ children }) => {
     return () => audioVideo.removeObserver(observer);
   }, [audioVideo, tileId]);
 
-  const value = useMemo(() => ({ tileId, isVideoEnabled, setIsVideoEnabled, toggleVideo, }), [
-    tileId,
-    isVideoEnabled,
-    setIsVideoEnabled,
-    toggleVideo,
-  ]);
+  const value = useMemo(
+    () => ({ tileId, isVideoEnabled, setIsVideoEnabled, toggleVideo }),
+    [tileId, isVideoEnabled, setIsVideoEnabled, toggleVideo]
+  );
 
   return <Context.Provider value={value}>{children}</Context.Provider>;
 };