+++ src/providers/MeetingProvider/MeetingManager.ts
@@ -2,41 +2,43 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {
+  AudioTransformDevice,
   AudioVideoFacade,
+  AudioVideoObserver,
   ConsoleLogger,
+  DefaultActiveSpeakerPolicy,
   DefaultDeviceController,
   DefaultMeetingSession,
+  Device,
+  EventAttributes,
+  EventName,
+  EventReporter,
+  isAudioTransformDevice,
+  isVideoTransformDevice,
+  Logger,
   LogLevel,
   MeetingSessionConfiguration,
-  DefaultActiveSpeakerPolicy,
+  MeetingSessionPOSTLogger,
   MeetingSessionStatus,
   MeetingSessionStatusCode,
-  AudioVideoObserver,
   MultiLogger,
-  MeetingSessionPOSTLogger,
-  EventReporter,
   VideoDownlinkBandwidthPolicy,
-  Logger,
-  EventName,
-  EventAttributes,
-  Device,
-  AudioTransformDevice,
-  isAudioTransformDevice,
+  VideoTransformDevice,
 } from 'amazon-chime-sdk-js';
 
+import { DeviceLabels, DeviceLabelTrigger, MeetingStatus } from '../../types';
 import {
   audioInputSelectionToDevice,
   supportsSetSinkId,
-  videoInputSelectionToDevice
+  videoInputSelectionToDevice,
 } from '../../utils/device-utils';
-import { DeviceLabels, DeviceLabelTrigger, MeetingStatus } from '../../types';
 import {
-  DevicePermissionStatus,
-  MeetingJoinData,
   AttendeeResponse,
+  DevicePermissionStatus,
   FullDeviceInfoType,
+  MeetingJoinData,
+  MeetingManagerConfig,
   PostLogConfig,
-  MeetingManagerConfig
 } from './types';
 
 function noOpDeviceLabelHook(): Promise<MediaStream> {
@@ -67,27 +69,42 @@ export class MeetingManager implements AudioVideoObserver {
 
   selectedAudioOutputDevice: string | null = null;
 
-  selectedAudioOutputDeviceObservers: ((deviceId: string | null) => void)[] = [];
+  selectedAudioOutputDeviceObservers: ((deviceId: string | null) => void)[] =
+    [];
 
   selectedAudioInputDevice: string | null = null;
 
-  selectedAudioInputTransformDevice: Device | AudioTransformDevice | null = null;
+  selectedAudioInputTransformDevice: Device | AudioTransformDevice | null =
+    null;
 
   selectedAudioInputDeviceObservers: ((deviceId: string | null) => void)[] = [];
 
-  selectedAudioInputTransformDeviceObservers: ((device: Device | AudioTransformDevice | null) => void)[] = [];
+  selectedAudioInputTransformDeviceObservers: ((
+    device: Device | AudioTransformDevice | null
+  ) => void)[] = [];
 
   selectAudioInputDeviceError: Error | null = null;
 
-  selectAudioInputDeviceErrorObservers: ((selectAudioInputDeviceError: Error | null) => void)[] = [];
+  selectAudioInputDeviceErrorObservers: ((
+    selectAudioInputDeviceError: Error | null
+  ) => void)[] = [];
 
   selectedVideoInputDevice: string | null = null;
+
+  selectedVideoInputTransformDevice: Device | VideoTransformDevice | null =
+    null;
+
+  selectedVideoInputTransformDeviceObservers: ((
+    device: Device | VideoTransformDevice | null
+  ) => void)[] = [];
 
   selectedVideoInputDeviceObservers: ((deviceId: string | null) => void)[] = [];
 
   selectVideoInputDeviceError: Error | null = null;
 
-  selectVideoInputDeviceErrorObservers: ((selectVideoInputDeviceError: Error | null) => void)[] = [];
+  selectVideoInputDeviceErrorObservers: ((
+    selectVideoInputDeviceError: Error | null
+  ) => void)[] = [];
 
   deviceLabelTriggerChangeObservers: (() => void)[] = [];
 
@@ -99,7 +116,8 @@ export class MeetingManager implements AudioVideoObserver {
 
   devicePermissionStatus = DevicePermissionStatus.UNSET;
 
-  devicePermissionsObservers: ((permission: DevicePermissionStatus) => void)[] = [];
+  devicePermissionsObservers: ((permission: DevicePermissionStatus) => void)[] =
+    [];
 
   activeSpeakerListener: ((activeSpeakers: string[]) => void) | null = null;
 
@@ -109,9 +127,8 @@ export class MeetingManager implements AudioVideoObserver {
 
   audioVideoCallbacks: ((audioVideo: AudioVideoFacade | null) => void)[] = [];
 
-  devicesUpdatedCallbacks: ((
-    fullDeviceInfo: FullDeviceInfoType
-  ) => void)[] = [];
+  devicesUpdatedCallbacks: ((fullDeviceInfo: FullDeviceInfoType) => void)[] =
+    [];
 
   // This variable will be deprecated in favor of `meetingManagerConfig`.
   // Please use `meetingManagerConfig` to use `MeetingManagerConfig` values.
@@ -135,7 +152,9 @@ export class MeetingManager implements AudioVideoObserver {
   // Please use `meetingManagerConfig` to use `MeetingManagerConfig` values.
   logger: Logger | undefined;
 
-  private meetingEventObserverSet = new Set<(name: EventName, attributes: EventAttributes) => void>();
+  private meetingEventObserverSet = new Set<
+    (name: EventName, attributes: EventAttributes) => void
+  >();
 
   private eventDidReceiveRef: AudioVideoObserver;
 
@@ -145,7 +164,7 @@ export class MeetingManager implements AudioVideoObserver {
       logger: configLogger,
       logLevel,
       postLogConfig,
-      videoDownlinkBandwidthPolicy
+      videoDownlinkBandwidthPolicy,
     } = this.meetingManagerConfig;
 
     // We are assigning the values from the `meetingManagerConfig` to preserve backward compatibility.
@@ -167,9 +186,11 @@ export class MeetingManager implements AudioVideoObserver {
       this.videoDownlinkBandwidthPolicy = videoDownlinkBandwidthPolicy;
     }
 
-    this.eventDidReceiveRef = { eventDidReceive: (name: EventName, attributes: EventAttributes) => {
-      this.publishEventDidReceiveUpdate(name, attributes);
-    }};
+    this.eventDidReceiveRef = {
+      eventDidReceive: (name: EventName, attributes: EventAttributes) => {
+        this.publishEventDidReceiveUpdate(name, attributes);
+      },
+    };
   }
 
   initializeMeetingManager(): void {
@@ -182,6 +203,7 @@ export class MeetingManager implements AudioVideoObserver {
     this.selectedAudioInputDevice = null;
     this.selectedAudioInputTransformDevice = null;
     this.selectedVideoInputDevice = null;
+    this.selectedVideoInputTransformDevice = null;
     this.selectAudioInputDeviceError = null;
     this.selectVideoInputDeviceError = null;
     this.audioInputDevices = [];
@@ -192,7 +214,13 @@ export class MeetingManager implements AudioVideoObserver {
     this.audioVideoObservers = {};
   }
 
-  async join({ meetingInfo, attendeeInfo, deviceLabels = DeviceLabels.AudioAndVideo, eventReporter, meetingManagerConfig }: MeetingJoinData) {
+  async join({
+    meetingInfo,
+    attendeeInfo,
+    deviceLabels = DeviceLabels.AudioAndVideo,
+    eventReporter,
+    meetingManagerConfig,
+  }: MeetingJoinData) {
     if (meetingManagerConfig) {
       this.meetingManagerConfig = meetingManagerConfig;
     }
@@ -244,14 +272,14 @@ export class MeetingManager implements AudioVideoObserver {
 
   async initializeMeetingSession(
     configuration: MeetingSessionConfiguration,
-    deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo,
+    deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo
   ): Promise<any> {
     const {
       simulcastEnabled,
       enableWebAudio,
       logger: configLogger,
       videoUplinkBandwidthPolicy,
-      videoDownlinkBandwidthPolicy
+      videoDownlinkBandwidthPolicy,
     } = this.meetingManagerConfig;
 
     // We are assigning the values from the `meetingManagerConfig` to preserve backward compatibility.
@@ -260,7 +288,9 @@ export class MeetingManager implements AudioVideoObserver {
       configuration.enableUnifiedPlanForChromiumBasedBrowsers = true;
       configuration.enableSimulcastForUnifiedPlanChromiumBasedBrowsers = true;
     }
-    const logger = configLogger ? configLogger : this.createLogger(configuration);
+    const logger = configLogger
+      ? configLogger
+      : this.createLogger(configuration);
 
     if (videoUplinkBandwidthPolicy) {
       configuration.videoUplinkBandwidthPolicy = videoUplinkBandwidthPolicy;
@@ -270,7 +300,9 @@ export class MeetingManager implements AudioVideoObserver {
       configuration.videoDownlinkBandwidthPolicy = videoDownlinkBandwidthPolicy;
     }
 
-    const deviceController = new DefaultDeviceController(logger, { enableWebAudio });
+    const deviceController = new DefaultDeviceController(logger, {
+      enableWebAudio,
+    });
     this.meetingSession = new DefaultMeetingSession(
       configuration,
       logger,
@@ -280,7 +312,7 @@ export class MeetingManager implements AudioVideoObserver {
 
     this.audioVideo = this.meetingSession.audioVideo;
     this.publishAudioVideo();
-    
+
     // When an attendee leaves, we remove AudioVideoObservers and nullify AudioVideoFacade and MeetingSession object.
     // This results into missing few meeting events triggered with audioVideoDidStop such as meetingEnded, meetingFailed and meetingStartFailed.
     // We may also loose audioInputUnselected and videoInputUnselected events as we choose null devices when an attendee leaves.
@@ -320,7 +352,9 @@ export class MeetingManager implements AudioVideoObserver {
   }
 
   audioVideoDidStart = () => {
-    console.log('[MeetingManager audioVideoDidStart] Meeting started successfully');
+    console.log(
+      '[MeetingManager audioVideoDidStart] Meeting started successfully'
+    );
     this.meetingStatus = MeetingStatus.Succeeded;
     this.publishMeetingStatus();
   };
@@ -328,20 +362,26 @@ export class MeetingManager implements AudioVideoObserver {
   audioVideoDidStop = (sessionStatus: MeetingSessionStatus) => {
     const sessionStatusCode = sessionStatus.statusCode();
     switch (sessionStatusCode) {
-      case MeetingSessionStatusCode.AudioCallEnded: 
-        console.log(`[MeetingManager audioVideoDidStop] Meeting ended for all: ${sessionStatusCode}`);
+      case MeetingSessionStatusCode.AudioCallEnded:
+        console.log(
+          `[MeetingManager audioVideoDidStop] Meeting ended for all: ${sessionStatusCode}`
+        );
         this.meetingStatus = MeetingStatus.Ended;
         this.publishMeetingStatus();
         this.leave();
         break;
       case MeetingSessionStatusCode.Left:
-        console.log(`[MeetingManager audioVideoDidStop] Left the meeting: ${sessionStatusCode}`);
+        console.log(
+          `[MeetingManager audioVideoDidStop] Left the meeting: ${sessionStatusCode}`
+        );
         this.meetingStatus = MeetingStatus.Left;
         this.publishMeetingStatus();
         this.leave();
         break;
       case MeetingSessionStatusCode.AudioJoinedFromAnotherDevice:
-        console.log(`[MeetingManager audioVideoDidStop] Meeting joined from another device: ${sessionStatusCode}`);
+        console.log(
+          `[MeetingManager audioVideoDidStop] Meeting joined from another device: ${sessionStatusCode}`
+        );
         this.meetingStatus = MeetingStatus.JoinedFromAnotherDevice;
         this.publishMeetingStatus();
         this.leave();
@@ -349,15 +389,21 @@ export class MeetingManager implements AudioVideoObserver {
       default:
         // The following status codes are Failures according to MeetingSessionStatus
         if (sessionStatus.isFailure()) {
-          console.log(`[MeetingManager audioVideoDidStop] Non-Terminal failure occured: ${sessionStatusCode}`);
+          console.log(
+            `[MeetingManager audioVideoDidStop] Non-Terminal failure occured: ${sessionStatusCode}`
+          );
           this.meetingStatus = MeetingStatus.Failed;
           this.publishMeetingStatus();
         } else if (sessionStatus.isTerminal()) {
-          console.log(`[MeetingManager audioVideoDidStop] Terminal failure occured: ${sessionStatusCode}`);
+          console.log(
+            `[MeetingManager audioVideoDidStop] Terminal failure occured: ${sessionStatusCode}`
+          );
           this.meetingStatus = MeetingStatus.TerminalFailure;
           this.publishMeetingStatus();
         }
-        console.log(`[MeetingManager audioVideoDidStop] session stopped with code ${sessionStatusCode}`);
+        console.log(
+          `[MeetingManager audioVideoDidStop] session stopped with code ${sessionStatusCode}`
+        );
         this.leave();
     }
 
@@ -373,7 +419,7 @@ export class MeetingManager implements AudioVideoObserver {
 
     this.audioVideoObservers = {
       audioVideoDidStart: this.audioVideoDidStart,
-      audioVideoDidStop: this.audioVideoDidStop
+      audioVideoDidStop: this.audioVideoDidStop,
     };
 
     this.audioVideo.addObserver(this.audioVideoObservers);
@@ -388,7 +434,9 @@ export class MeetingManager implements AudioVideoObserver {
       (await this.audioVideo?.listAudioOutputDevices()) || [];
   }
 
-  setupDeviceLabelTrigger(deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo): void {
+  setupDeviceLabelTrigger(
+    deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo
+  ): void {
     let callback: DeviceLabelTrigger;
 
     if (typeof deviceLabels === 'function') {
@@ -416,7 +464,9 @@ export class MeetingManager implements AudioVideoObserver {
         this.publishDevicePermissionStatus();
         try {
           const devices = await navigator.mediaDevices.enumerateDevices();
-          const hasVideoInput = devices.some(value => value.kind === 'videoinput');
+          const hasVideoInput = devices.some(
+            (value) => value.kind === 'videoinput'
+          );
 
           const stream = await navigator.mediaDevices.getUserMedia({
             audio: constraints.audio,
@@ -439,22 +489,26 @@ export class MeetingManager implements AudioVideoObserver {
   }
 
   setupActiveSpeakerDetection(): void {
-    const activeSpeakerPolicy  = this.meetingManagerConfig.activeSpeakerPolicy;
+    const activeSpeakerPolicy = this.meetingManagerConfig.activeSpeakerPolicy;
 
     this.publishActiveSpeaker();
 
     this.activeSpeakerListener = (activeSpeakers: string[]) => {
       this.activeSpeakers = activeSpeakers;
-      this.activeSpeakerCallbacks.forEach(cb => cb(activeSpeakers));
+      this.activeSpeakerCallbacks.forEach((cb) => cb(activeSpeakers));
     };
 
     this.audioVideo?.subscribeToActiveSpeakerDetector(
-      activeSpeakerPolicy ? activeSpeakerPolicy : new DefaultActiveSpeakerPolicy(),
+      activeSpeakerPolicy
+        ? activeSpeakerPolicy
+        : new DefaultActiveSpeakerPolicy(),
       this.activeSpeakerListener
     );
   }
 
-  async listAndSelectDevices(deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo): Promise<void> {
+  async listAndSelectDevices(
+    deviceLabels: DeviceLabels | DeviceLabelTrigger = DeviceLabels.AudioAndVideo
+  ): Promise<void> {
     await this.updateDeviceLists();
 
     // If `deviceLabels` is of `DeviceLabelTrigger` type, no device will be selected.
@@ -486,6 +540,7 @@ export class MeetingManager implements AudioVideoObserver {
       this.audioInputDevices.length
     ) {
       this.selectedAudioInputDevice = this.audioInputDevices[0].deviceId;
+      this.selectedAudioInputTransformDevice = this.selectedAudioInputDevice;
       try {
         await this.audioVideo?.chooseAudioInputDevice(
           this.audioInputDevices[0].deviceId
@@ -494,6 +549,7 @@ export class MeetingManager implements AudioVideoObserver {
         console.error(`Error in selecting audio input device - ${error}`);
       }
       this.publishSelectedAudioInputDevice();
+      this.publishSelectedAudioInputTransformDevice();
     }
     if (
       isAudioDeviceRequested &&
@@ -520,11 +576,15 @@ export class MeetingManager implements AudioVideoObserver {
       this.videoInputDevices.length
     ) {
       this.selectedVideoInputDevice = this.videoInputDevices[0].deviceId;
+      this.selectedVideoInputTransformDevice = this.selectedVideoInputDevice;
       this.publishSelectedVideoInputDevice();
+      this.publishSelectedVideoInputTransformDevice();
     }
   }
 
-  selectAudioInputDevice = async (device: Device | AudioTransformDevice): Promise<void> => {
+  selectAudioInputDevice = async (
+    device: Device | AudioTransformDevice
+  ): Promise<void> => {
     let receivedDevice = device;
     if (typeof device === 'string') {
       receivedDevice = audioInputSelectionToDevice(device);
@@ -555,17 +615,13 @@ export class MeetingManager implements AudioVideoObserver {
         innerDevice = device;
       }
 
-      if (innerDevice === null) {
-        this.selectedAudioInputDevice = null;
-      } else if (typeof innerDevice === 'string') {
-        this.selectedAudioInputDevice = innerDevice;
-      } else if (innerDevice instanceof MediaStream) {
-        this.selectedAudioInputDevice = innerDevice.id;
-      } else {
-        const deviceId = this.getDeviceID(innerDevice);
+      const deviceId =
+        DefaultDeviceController.getIntrinsicDeviceId(innerDevice);
+      if (typeof deviceId === 'string') {
         this.selectedAudioInputDevice = deviceId;
+      } else if (Array.isArray(deviceId) && deviceId[0]) {
+        this.selectedAudioInputDevice = deviceId[0];
       }
-
       this.selectedAudioInputTransformDevice = device;
     }
     this.publishSelectedAudioInputDevice();
@@ -573,39 +629,23 @@ export class MeetingManager implements AudioVideoObserver {
     this.publishSelectedAudioInputTransformDevice();
   };
 
-  private getDeviceID = (innerDevice: MediaTrackConstraints | MediaTrackConstraintSet): string => {
-    if (Array.isArray(innerDevice)) {
-      innerDevice = innerDevice[0];
-    }
-
-    const deviceId = innerDevice.deviceId;
-    if (typeof (deviceId) === 'string') {
-      return deviceId;
-    }
-    if (Array.isArray(deviceId)) {
-      return deviceId[0];
-    }
-    if (Array.isArray(deviceId?.exact) && deviceId?.exact) {
-      return deviceId.exact[0];
-    }
-    if (typeof (deviceId?.exact) === 'string') {
-      return deviceId.exact;
-    }
-    return '';
-  }
-
   selectAudioOutputDevice = async (deviceId: string): Promise<void> => {
     try {
       await this.audioVideo?.chooseAudioOutputDevice(deviceId);
       this.selectedAudioOutputDevice = deviceId;
       this.publishSelectedAudioOutputDevice();
     } catch (error) {
-      console.error(`Error setting audio output - ${error}`);
+      console.error(`Error setting audio output`, error);
     }
   };
 
-  selectVideoInputDevice = async (deviceId: string): Promise<void> => {
-    const receivedDevice = videoInputSelectionToDevice(deviceId);
+  selectVideoInputDevice = async (
+    device: Device | VideoTransformDevice
+  ): Promise<void> => {
+    let receivedDevice = device;
+    if (typeof device === 'string') {
+      receivedDevice = videoInputSelectionToDevice(device);
+    }
     if (receivedDevice === null) {
       try {
         await this.audioVideo?.chooseVideoInputDevice(null);
@@ -615,6 +655,7 @@ export class MeetingManager implements AudioVideoObserver {
         console.error('Failed to choose video input device.', error);
       }
       this.selectedVideoInputDevice = null;
+      this.selectedVideoInputTransformDevice = null;
     } else {
       try {
         await this.audioVideo?.chooseVideoInputDevice(receivedDevice);
@@ -623,10 +664,25 @@ export class MeetingManager implements AudioVideoObserver {
         this.selectVideoInputDeviceError = error;
         console.error('Failed to choose video input device.', error);
       }
-      this.selectedVideoInputDevice = deviceId;
+      let innerDevice = null;
+      if (isVideoTransformDevice(device)) {
+        innerDevice = await device.intrinsicDevice();
+      } else {
+        innerDevice = device;
+      }
+
+      const deviceId =
+        DefaultDeviceController.getIntrinsicDeviceId(innerDevice);
+      if (typeof deviceId === 'string') {
+        this.selectedVideoInputDevice = deviceId;
+      } else if (Array.isArray(deviceId) && deviceId[0]) {
+        this.selectedVideoInputDevice = deviceId[0];
+      }
+      this.selectedVideoInputTransformDevice = device;
     }
     this.publishSelectedVideoInputDevice();
     this.publishSelectVideoInputDeviceError();
+    this.publishSelectedVideoInputTransformDevice();
   };
 
   invokeDeviceProvider = (deviceLabels: DeviceLabels) => {
@@ -650,12 +706,12 @@ export class MeetingManager implements AudioVideoObserver {
     callbackToRemove: (av: AudioVideoFacade | null) => void
   ): void => {
     this.audioVideoCallbacks = this.audioVideoCallbacks.filter(
-      callback => callback !== callbackToRemove
+      (callback) => callback !== callbackToRemove
     );
   };
 
   publishAudioVideo = () => {
-    this.audioVideoCallbacks.forEach(callback => {
+    this.audioVideoCallbacks.forEach((callback) => {
       callback(this.audioVideo);
     });
   };
@@ -671,12 +727,12 @@ export class MeetingManager implements AudioVideoObserver {
     callbackToRemove: (activeSpeakers: string[]) => void
   ): void => {
     this.activeSpeakerCallbacks = this.activeSpeakerCallbacks.filter(
-      callback => callback !== callbackToRemove
+      (callback) => callback !== callbackToRemove
     );
   };
 
   publishActiveSpeaker = () => {
-    this.activeSpeakerCallbacks.forEach(callback => {
+    this.activeSpeakerCallbacks.forEach((callback) => {
       callback(this.activeSpeakers);
     });
   };
@@ -691,7 +747,7 @@ export class MeetingManager implements AudioVideoObserver {
     callbackToRemove: (permission: DevicePermissionStatus) => void
   ): void => {
     this.devicePermissionsObservers = this.devicePermissionsObservers.filter(
-      callback => callback !== callbackToRemove
+      (callback) => callback !== callbackToRemove
     );
   };
 
@@ -710,14 +766,36 @@ export class MeetingManager implements AudioVideoObserver {
   unsubscribeFromSelectedVideoInputDevice = (
     callbackToRemove: (deviceId: string | null) => void
   ): void => {
-    this.selectedVideoInputDeviceObservers = this.selectedVideoInputDeviceObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectedVideoInputDeviceObservers =
+      this.selectedVideoInputDeviceObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishSelectedVideoInputDevice = (): void => {
     for (const observer of this.selectedVideoInputDeviceObservers) {
       observer(this.selectedVideoInputDevice);
+    }
+  };
+
+  subscribeToSelectedVideoInputTransformDevice = (
+    callback: (device: Device | VideoTransformDevice) => void
+  ): void => {
+    this.selectedVideoInputTransformDeviceObservers.push(callback);
+  };
+
+  unsubscribeFromSelectedVideoInputTranformDevice = (
+    callbackToRemove: (device: Device | VideoTransformDevice) => void
+  ): void => {
+    this.selectedVideoInputTransformDeviceObservers =
+      this.selectedVideoInputTransformDeviceObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
+  };
+
+  private publishSelectedVideoInputTransformDevice = (): void => {
+    for (const observer of this.selectedVideoInputTransformDeviceObservers) {
+      observer(this.selectedVideoInputTransformDevice);
     }
   };
 
@@ -736,17 +814,19 @@ export class MeetingManager implements AudioVideoObserver {
   unsubscribeFromSelectedAudioInputDevice = (
     callbackToRemove: (deviceId: string | null) => void
   ): void => {
-    this.selectedAudioInputDeviceObservers = this.selectedAudioInputDeviceObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectedAudioInputDeviceObservers =
+      this.selectedAudioInputDeviceObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   unsubscribeFromSelectedAudioInputTransformDevice = (
     callbackToRemove: (device: Device | AudioTransformDevice | null) => void
   ): void => {
-    this.selectedAudioInputTransformDeviceObservers = this.selectedAudioInputTransformDeviceObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectedAudioInputTransformDeviceObservers =
+      this.selectedAudioInputTransformDeviceObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishSelectedAudioInputDevice = (): void => {
@@ -770,9 +850,10 @@ export class MeetingManager implements AudioVideoObserver {
   unsubscribeFromSelectedAudioOutputDevice = (
     callbackToRemove: (deviceId: string | null) => void
   ): void => {
-    this.selectedAudioOutputDeviceObservers = this.selectedAudioOutputDeviceObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectedAudioOutputDeviceObservers =
+      this.selectedAudioOutputDeviceObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishSelectedAudioOutputDevice = (): void => {
@@ -792,12 +873,12 @@ export class MeetingManager implements AudioVideoObserver {
     callbackToRemove: (meetingStatus: MeetingStatus) => void
   ): void => {
     this.meetingStatusObservers = this.meetingStatusObservers.filter(
-      callback => callback !== callbackToRemove
+      (callback) => callback !== callbackToRemove
     );
   };
 
   private publishMeetingStatus = () => {
-    this.meetingStatusObservers.forEach(callback => {
+    this.meetingStatusObservers.forEach((callback) => {
       callback(this.meetingStatus);
     });
   };
@@ -811,9 +892,10 @@ export class MeetingManager implements AudioVideoObserver {
   unsubscribeFromSelectAudioInputDeviceError = (
     callbackToRemove: (errorMessage: Error | null) => void
   ): void => {
-    this.selectAudioInputDeviceErrorObservers = this.selectAudioInputDeviceErrorObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectAudioInputDeviceErrorObservers =
+      this.selectAudioInputDeviceErrorObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishSelectAudioInputDeviceError = (): void => {
@@ -831,9 +913,10 @@ export class MeetingManager implements AudioVideoObserver {
   unsubscribeFromSelectVideoInputDeviceError = (
     callbackToRemove: (errorMessage: Error | null) => void
   ): void => {
-    this.selectVideoInputDeviceErrorObservers = this.selectVideoInputDeviceErrorObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.selectVideoInputDeviceErrorObservers =
+      this.selectVideoInputDeviceErrorObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishSelectVideoInputDeviceError = (): void => {
@@ -842,18 +925,17 @@ export class MeetingManager implements AudioVideoObserver {
     }
   };
 
-  subscribeToDeviceLabelTriggerChange = (
-    callback: () => void
-  ): void => {
+  subscribeToDeviceLabelTriggerChange = (callback: () => void): void => {
     this.deviceLabelTriggerChangeObservers.push(callback);
   };
 
   unsubscribeFromDeviceLabelTriggerChange = (
     callbackToRemove: () => void
   ): void => {
-    this.deviceLabelTriggerChangeObservers = this.deviceLabelTriggerChangeObservers.filter(
-      callback => callback !== callbackToRemove
-    );
+    this.deviceLabelTriggerChangeObservers =
+      this.deviceLabelTriggerChangeObservers.filter(
+        (callback) => callback !== callbackToRemove
+      );
   };
 
   private publishDeviceLabelTriggerChange = (): void => {
@@ -862,18 +944,26 @@ export class MeetingManager implements AudioVideoObserver {
     }
   };
 
-  subscribeToEventDidReceive = (callback: (name: EventName, attributes: EventAttributes) => void): void => {
+  subscribeToEventDidReceive = (
+    callback: (name: EventName, attributes: EventAttributes) => void
+  ): void => {
     this.meetingEventObserverSet.add(callback);
   };
 
-  unsubscribeFromEventDidReceive = (callbackToRemove: (name: EventName, attributes: EventAttributes) => void): void => {
+  unsubscribeFromEventDidReceive = (
+    callbackToRemove: (name: EventName, attributes: EventAttributes) => void
+  ): void => {
     this.meetingEventObserverSet.delete(callbackToRemove);
   };
 
-  private publishEventDidReceiveUpdate = (name: EventName, attributes: EventAttributes) => {
-    this.meetingEventObserverSet.forEach((callback) => callback(name, attributes));
+  private publishEventDidReceiveUpdate = (
+    name: EventName,
+    attributes: EventAttributes
+  ) => {
+    this.meetingEventObserverSet.forEach((callback) =>
+      callback(name, attributes)
+    );
   };
-
 }
 
 export default MeetingManager;